Extract common gateway methods to module
This is beneficial for sharing code as well as for unifying behaviour.
The existing HPP Gateway's behaviour has preference.
